requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
argparse=1.4.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.7=py312he558020_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
certifi=2024.2.2=py312haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py312h59b6b97_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py312hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
eigen=3.4.0=h59b6b97_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=hd77b12b_0
ffmpeg=4.2.2=he774522_0
filelock=3.13.1=py312haa95532_0
fonttools=4.51.0=py312h2bbff1b_0
freetype=2.12.1=ha860e81_0
fsspec=2024.6.0=pypi_0
giflib=5.2.1=h8cc25b3_3
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
grpcio=1.64.1=pypi_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.12.1=h51c971a_3
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.7=py312haa95532_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2021.4.0=pypi_0
ipykernel=6.29.3=pyha63f2e9_0
ipython=8.24.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=py312haa95532_0
joblib=1.4.0=py312haa95532_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.5.0=py312haa95532_0
kiwisolver=1.4.4=py312hd77b12b_0
krb5=1.21.2=heb0366b_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=12.0.0=default_h627e005_2
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libglib=2.78.4=ha17d25a_0
libiconv=1.16=h2bbff1b_3
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_2
libxslt=1.1.37=h2bbff1b_1
lightning-utilities=0.11.2=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markdown=3.6=pypi_0
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mlxtend=0.23.1=pypi_0
mpmath=1.3.0=py312haa95532_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=py312haa95532_0
nibabel=5.2.1=pypi_0
numexpr=2.8.7=py312h96b7d27_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
opencv=4.6.0=py312h14b9924_5
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.1.4=pypi_0
openssl=3.3.0=h2466b09_3
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.42=h0ff8eda_1
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h2bbff1b_0
pip=24.0=py312haa95532_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=py312haa95532_1
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py312h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pybind11-abi=5=hd3eb1b0_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.0.9=py312haa95532_0
pyqt=5.15.10=py312hd77b12b_0
pyqt5-sip=12.13.0=py312h2bbff1b_0
pysocks=1.7.1=py312haa95532_0
python=3.12.3=h1d929f7_1
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.3.0=py3.12_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py312haa95532_0
pywin32=305=py312h2bbff1b_0
pyyaml=6.0.1=py312h2bbff1b_0
pyzmq=25.1.2=py312hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
requests=2.31.0=py312haa95532_1
scikit-learn=1.4.2=py312hc7c4135_1
scipy=1.13.0=py312hbb039d4_0
seaborn=0.12.2=py312haa95532_0
setuptools=69.5.1=py312haa95532_0
sip=6.7.12=py312hd77b12b_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py312haa95532_0
tbb=2021.12.0=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.14=h0416ee5_0
torchaudio=2.3.0=pypi_0
torchinfo=1.8.0=pypi_0
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.18.0=pypi_0
tornado=6.3.3=py312h2bbff1b_0
tqdm=4.66.4=py312hfc267ef_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024a=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.1.0=py312h2bbff1b_0
urllib3=2.2.1=py312haa95532_0
utility=1.0=pypi_0
vc=14.2=h2eaa2aa_1
vc14_runtime=14.38.33135=h835141b_20
vs2015_runtime=14.38.33135=h22015db_20
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zeromq=4.3.5=he1f189c_4
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
